pyproject.toml
[tool.poetry]
name = "django-structure"
version = "0.1.0"
description = ""
authors = ["palas1k <liserg008@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-filter = "^23.4"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
